STREAMLIT/airquality/app/requirements.txt
# Requirements pour l'application de qualité de l'air géolocalisée
openmeteo-requests>=1.1.0
requests-cache>=1.0.0
retry-requests>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
folium>=0.15.0
numpy>=1.24.0
plotly>=5.17.0
dash>=2.14.0
geopy>=2.4.0
scipy>=1.11.0
scikit-learn>=1.3.0
streamlit>=1.40.0
streamlit-folium==0.17.0
bcrypt>=4.0.0
PyJWT>=2.8.0
mailjet-rest>=1.3.4
python-dotenv>=1.0.0

# PostgreSQL + Prisma
prisma>=0.11.0
psycopg2-binary>=2.9.9

# Gestion Session & Cookies
extra-streamlit-components>=0.1.0
redis>=5.0.0